requirements.txt
smtplib
email
random
string
datetime
python-dotenv
tkinter
